requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg[binary]==3.2.9
aiosqlite==0.21.0
celery[redis]==5.5.3
redis==5.2.1
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
openpyxl==3.1.5
pandas==2.3.1
reportlab==4.4.3
boto3==1.40.11
python-multipart==0.0.20
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1
